+++ vidigi/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __author__ = "Sammi Rosser"
 
-from . import animation, ciw, prep, utils, logging, resources
+from . import animation, ciw, prep, utils, logging, resources, process_mapping